[Feature] When level points are paid, connect when green label points are paid [Issue no]  #140

+++ src/main/java/com/impacus/maketplace/common/enumType/user/UserLevel.java
@@ -9,18 +9,18 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public enum UserLevel {
 
-    NONE("등급 없음", 0, 0, 0, 0, 0),
-    BRONZE("BRONZE", 1,0,100000, 0, 1.0),
-    ROOKIE("ROOKIE", 2,100001,300000, 10000, 1.5),
-    SILVER("SILVER", 3, 300001, 500000, 50000, 2.0),
-    GOLD("GOLD", 4,500001,900000, 90000, 2.5),
-    ECO_VIP("ECO_VIP", 5, 900001, Integer.MAX_VALUE, 50000, 3.0);
+    NONE("등급 없음", 0, 0, 0, 0L, 0),
+    BRONZE("BRONZE", 1, 0, 100000, 0L, 1.0),
+    ROOKIE("ROOKIE", 2, 100001, 300000, 10000L, 1.5),
+    SILVER("SILVER", 3, 300001, 500000, 50000L, 2.0),
+    GOLD("GOLD", 4, 500001, 900000, 90000L, 2.5),
+    ECO_VIP("ECO_VIP", 5, 900001, Integer.MAX_VALUE, 50000L, 3.0);
 
     private final String value;
     private final int level;
     private final int minScore;
     private final int maxScore;
-    private final int celebrationPoint;
+    private final Long celebrationPoint;
     private final double reserveRate;
 
     /**
